requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.6.15.2
charset-normalizer==2.1.1
click==8.1.3
dbf-light==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
google-api-core==2.10.0
google-api-python-client==2.61.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
npm==0.1.1
numpy==1.23.3
oauthlib==3.2.1
optional-django==0.1.0
pandas==1.4.4
protobuf==4.21.6
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycbrf==1.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
